% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{assert_apikey}
\alias{assert_apikey}
\title{Checks whether apikey is correctly set}
\usage{
assert_apikey(key)
}
\arguments{
\item{key}{airly apikey}
}
\description{
Checks whether apikey is correctly set
}
