% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/assertion.R
\name{assert_coordinates}
\alias{assert_coordinates}
\title{Checks whether apikey is correctly set}
\usage{
assert_coordinates(lat, lng)
}
\arguments{
\item{lat}{latitude as decimal degree}

\item{lng}{longitude as decimal degree}
}
\description{
Checks whether apikey is correctly set
}
