% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{build_forecast_df}
\alias{build_forecast_df}
\title{Creates object containing information about history data for given API response}
\usage{
build_forecast_df(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
tibble representing a airly_measurement with time, measures and indexes fields
}
\description{
Creates object containing information about history data for given API response
}
