% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{create_airly_index}
\alias{create_airly_index}
\title{Creates an object representing Airly index}
\usage{
create_airly_index(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
object representing a airly_index
}
\description{
Creates an object representing Airly index
}
