% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_limit.R
\name{create_airly_limit}
\alias{create_airly_limit}
\title{Creates an object representing Airly limit}
\usage{
create_airly_limit()
}
\value{
object representing an airly_limit
}
\description{
Creates an object representing Airly limit
}
