% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{create_airly_measurement}
\alias{create_airly_measurement}
\title{Creates an object representing Airly measurement}
\usage{
create_airly_measurement(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
object representing a airly_measurement
}
\description{
Creates an object representing Airly measurement
}
