% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{.send_request}
\alias{.send_request}
\title{Sends a request to the specified url and retrieves it's content.}
\usage{
.send_request(request_url, apikey, query = NULL)
}
\arguments{
\item{request_url}{url to be used}

\item{apikey}{airly apikey}

\item{query}{Default value is NULL. Optional argument if you want to add query to request}
}
\value{
parsed content of the response object
}
\description{
Sends a request to the specified url and retrieves it's content.
}
