% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_indexes}
\alias{get_indexes}
\title{Get Airly available indexes}
\usage{
get_indexes()
}
\value{
object of airly_meta class
}
\description{
Endpoint returns a list of all the index types supported in the API along with lists of levels defined per each index type.
}
\examples{
\donttest{
 get_indexes()
}

}
