% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_installation_measurements}
\alias{get_installation_measurements}
\title{Get Airly measurements for any geographical location given installation id}
\usage{
get_installation_measurements(id)
}
\arguments{
\item{id}{integer, installation identifier}
}
\value{
object of airly_measurements class
}
\description{
Endpoint returns measurements for concrete installation given by installation Id
}
\examples{
\donttest{
 get_installation_measurements(8077)
}

}
