% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_measurements_info}
\alias{get_measurements_info}
\title{Get measures used in Airly}
\usage{
get_measurements_info()
}
\value{
data.frame with measure names and units
}
\description{
Endpoint returns list of all the measurement types supported in the API along with their names and units.
}
\examples{
\donttest{
 get_measurements_info()
}

}
