% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_nearest_installations}
\alias{get_nearest_installations}
\title{Get Airly nearest installations to given point}
\usage{
get_nearest_installations(lat, lng, max_distance = NULL, max_results = NULL)
}
\arguments{
\item{lat}{latitude as decimal degree}

\item{lng}{longitude as decimal degree}

\item{max_distance}{default value 3.0. All the returned installations must be located within this limit from the given point (in km). Negative value means no limit}

\item{max_results}{default value 1. Maximum number of installations to return. Negative value means no limit}
}
\value{
data.frame of airly_location items
}
\description{
Endpoint returns list of installations which are closest to a given point, sorted by distance to that point.
}
\examples{
\donttest{
get_nearest_installations(50.11670, 19.91429, max_distance = 20)
}

}
