% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_nearest_measurements}
\alias{get_nearest_measurements}
\title{Get Airly nearest measurements to given point}
\usage{
get_nearest_measurements(lat, lng, max_distance = NULL)
}
\arguments{
\item{lat}{latitude as decimal degree}

\item{lng}{longitude as decimal degree}

\item{max_distance}{default value 3.0. All the returned installations must be located within this limit from the given point (in km). Negative value means no limit}
}
\value{
data.frame of airly_measurements items
}
\description{
Endpoint returns measurements for an installation closest to a given location
}
\examples{
\donttest{
get_nearest_measurements(50.11670, 19.91429, max_distance = 10)
}

}
