% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_point_measurements}
\alias{get_point_measurements}
\title{Get Airly measurements for any geographical location}
\usage{
get_point_measurements(lat, lng)
}
\arguments{
\item{lat}{latitude as decimal degree}

\item{lng}{longitude as decimal degree}
}
\value{
object of airly_measurements class
}
\description{
Endpoint returns measurements for any geographical location
}
\examples{
\donttest{
get_point_measurements(50.11670, 19.91429)
}

}
