% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_api_response.R
\name{is_airly_api_response}
\alias{is_airly_api_response}
\title{Checks whether the given object is of the class
airly_api_response}
\usage{
is_airly_api_response(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_api_response}
}
\value{
TRUE if the object is of the class airly_api_response
}
\description{
Checks whether the given object is of the class
airly_api_response
}
