% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{is_airly_index}
\alias{is_airly_index}
\title{Checks whether the given object is of the class
airly_index}
\usage{
is_airly_index(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_index}
}
\value{
TRUE if the object is of the class airly_index
}
\description{
Checks whether the given object is of the class
airly_index
}
