% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_limit.R
\name{is_airly_limit}
\alias{is_airly_limit}
\title{Checks whether the given object is of the class
airly_limit}
\usage{
is_airly_limit(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_limit}
}
\value{
TRUE if the object is of the class airly_limit
}
\description{
Checks whether the given object is of the class
airly_limit
}
