% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_location.R
\name{is_airly_location}
\alias{is_airly_location}
\title{Checks whether the given object is of the class
airly_location}
\usage{
is_airly_location(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_location}
}
\value{
TRUE if the object is of the class airly_location
}
\description{
Checks whether the given object is of the class
airly_location
}
