% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{is_airly_measurement}
\alias{is_airly_measurement}
\title{Checks whether the given object is of the class
airly_measurement}
\usage{
is_airly_measurement(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_measurement}
}
\value{
TRUE if the object is of the class airly_measurement
}
\description{
Checks whether the given object is of the class
airly_measurement
}
