% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_meta.R
\name{is_airly_meta}
\alias{is_airly_meta}
\title{Checks whether the given object is of the class
airly_meta}
\usage{
is_airly_meta(x)
}
\arguments{
\item{x}{object to test if it is of the class airly_meta}
}
\value{
TRUE if the object is of the class airly_meta
}
\description{
Checks whether the given object is of the class
airly_meta
}
