% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{print.airly_index}
\alias{print.airly_index}
\title{Print for "airly_index" type objects}
\usage{
\method{print}{airly_index}(x, ...)
}
\arguments{
\item{x}{"airly_index" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "airly_index" type objects
}
