% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_limit.R
\name{print.airly_limit}
\alias{print.airly_limit}
\title{Print for "airly_limit" type objects}
\usage{
\method{print}{airly_limit}(x, ...)
}
\arguments{
\item{x}{"airly_limit" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "airly_limit" type objects
}
