% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{print.airly_location}
\alias{print.airly_location}
\title{Print for "airly_location" type objects}
\usage{
\method{print}{airly_location}(x, ...)
}
\arguments{
\item{x}{"airly_location" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "airly_location" type objects
}
