% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{print.airly_measurement}
\alias{print.airly_measurement}
\title{Print for "airly_measurement" type objects}
\usage{
\method{print}{airly_measurement}(x, ...)
}
\arguments{
\item{x}{"airly_measurement" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "airly_measurement" type objects
}
