% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_meta.R
\name{print.airly_meta}
\alias{print.airly_meta}
\title{Print for "airly_meta" type objects}
\usage{
\method{print}{airly_meta}(x, ...)
}
\arguments{
\item{x}{"airly_meta" type list}

\item{...}{further arguments passed to or from other methods}
}
\description{
Print for "airly_meta" type objects
}
