% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_null}
\alias{replace_null}
\title{Replaces NULL with NA for nested lists. Useful when NULL value
leads to error while object casting}
\usage{
replace_null(x)
}
\arguments{
\item{x}{nested list}
}
\value{
same list with NULL replaced with NA
}
\description{
Replaces NULL with NA for nested lists. Useful when NULL value
leads to error while object casting
}
