% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{set_apikey}
\alias{set_apikey}
\title{Set Airly apikey}
\usage{
set_apikey(key)
}
\arguments{
\item{key}{string. Get your api key \url{https://developer.airly.eu/}}
}
\description{
On a free plan, API consumer is required to use our API only in non-commercial projects.
More details are available in under \url{https://airly.eu/docs/tos-en.pdf}.
}
\examples{
\donttest{
set_apikey("abctest")
}

}
