% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_api_response.R
\name{validate_airly_api_response}
\alias{validate_airly_api_response}
\title{Checks if the given response is not empty and that
it did not return an error http code.}
\usage{
validate_airly_api_response(airly_api_response)
}
\arguments{
\item{airly_api_response}{airly_api_response object to be checked}
}
\description{
Checks if the given response is not empty and that
it did not return an error http code.
}
