% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{validate_airly_index}
\alias{validate_airly_index}
\title{Checks whether the given object is correctly defined
airly_index class}
\usage{
validate_airly_index(airly_index)
}
\arguments{
\item{airly_index}{object of the class airly_index}
}
\description{
Checks whether the given object is correctly defined
airly_index class
}
