% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_limit.R
\name{validate_airly_limit}
\alias{validate_airly_limit}
\title{Checks whether the given object is correctly defined
airly_limit class}
\usage{
validate_airly_limit(airly_limit)
}
\arguments{
\item{airly_limit}{object of the class airly_limit}
}
\description{
Checks whether the given object is correctly defined
airly_limit class
}
