% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_location.R
\name{validate_airly_location}
\alias{validate_airly_location}
\title{Checks whether the given object is correctly defined
airly_location class}
\usage{
validate_airly_location(airly_location)
}
\arguments{
\item{airly_location}{tibble airly_location}
}
\description{
Checks whether the given object is correctly defined
airly_location class
}
