% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{validate_airly_measurement}
\alias{validate_airly_measurement}
\title{Checks whether the given object is correctly defined
airly_measurement class}
\usage{
validate_airly_measurement(airly_measurement)
}
\arguments{
\item{airly_measurement}{object of the class airly_measurement}
}
\description{
Checks whether the given object is correctly defined
airly_measurement class
}
