% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_meta.R
\name{validate_airly_meta}
\alias{validate_airly_meta}
\title{Checks whether the given object is correctly correctly defined}
\usage{
validate_airly_meta(airly_meta)
}
\arguments{
\item{airly_meta}{object of the class airly_meta}
}
\description{
Checks whether the given object is correctly correctly defined
}
