% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_measurement.R
\name{build_forecast_df}
\alias{build_forecast_df}
\title{Creates an object representing Airly measurement}
\usage{
build_forecast_df(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
object representing a airly_measurement
}
\description{
Creates an object representing Airly measurement
}
