% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_api_response.R
\name{create_airly_api_response}
\alias{create_airly_api_response}
\title{Creates an object representing a response from the
Airly API}
\usage{
create_airly_api_response(response)
}
\arguments{
\item{response}{response object}
}
\value{
object representing a response from the
Airly API
}
\description{
Creates an object representing a response from the
Airly API
}
