% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_index.R
\name{create_airly_location}
\alias{create_airly_location}
\title{Creates an object representing Airly location}
\usage{
create_airly_location(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
object representing a airly_location
}
\description{
Creates an object representing Airly location
}
