% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airly_meta.R
\name{create_airly_meta}
\alias{create_airly_meta}
\title{Creates an object representing Airly meta}
\usage{
create_airly_meta(item)
}
\arguments{
\item{item}{list returned by Airly API}
}
\value{
object representing an airly_meta
}
\description{
Creates an object representing Airly meta
}
