% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{get_installation_by_id}
\alias{get_installation_by_id}
\title{Get Airly installation by id}
\usage{
get_installation_by_id(id)
}
\arguments{
\item{id}{integer}
}
\value{
airly_index item
}
\description{
Endpoint returns single installation metadata, given by id
}
\details{

}
\examples{
\donttest{
get_installation_by_id(8077)
}

}
